+++ pom.xml
@@ -18,6 +18,7 @@
         <testng.version>7.8.0</testng.version>
         <selenium-java.version>4.9.1</selenium-java.version>
         <assertj-core.version>3.24.2</assertj-core.version>
+        <javafaker.version>1.0.2</javafaker.version>
 
         <maven-checkstyle-plugin.version>3.2.0</maven-checkstyle-plugin.version>
     </properties>
@@ -48,6 +49,13 @@
             <version>${assertj-core.version}</version>
             <scope>test</scope>
         </dependency>
+
+        <dependency>
+            <groupId>com.github.javafaker</groupId>
+            <artifactId>javafaker</artifactId>
+            <version>${javafaker.version}</version>
+        </dependency>
+
     </dependencies>
 
     <build>
