+++ src/main/java/ru/levelp/at/lesson0507/selenium/page/objects/component/elements/BaseComponent.java
@@ -1,0 +1,19 @@
+package ru.levelp.at.lesson0507.selenium.page.objects.component.elements;
+
+import java.time.Duration;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.support.PageFactory;
+import org.openqa.selenium.support.ui.WebDriverWait;
+
+public abstract class BaseComponent {
+
+    protected final WebDriver driver;
+    protected final WebDriverWait wait;
+
+    protected BaseComponent(WebDriver driver) {
+        this.driver = driver;
+        wait = new WebDriverWait(driver, Duration.ofSeconds(10));
+
+        PageFactory.initElements(driver, this);
+    }
+}
