+++ src/main/java/ru/levelp/at/lesson0507/selenium/page/objects/component/elements/LoginComponent.java
@@ -1,0 +1,10 @@
+package ru.levelp.at.lesson0507.selenium.page.objects.component.elements;
+
+import org.openqa.selenium.WebDriver;
+
+public class LoginComponent extends BaseComponent {
+
+    public LoginComponent(WebDriver driver) {
+        super(driver);
+    }
+}
