+++ src/main/java/ru/levelp/at/lesson0507/selenium/page/objects/component/elements/RegistrationComponent.java
@@ -1,0 +1,48 @@
+package ru.levelp.at.lesson0507.selenium.page.objects.component.elements;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.FindBy;
+import org.openqa.selenium.support.ui.ExpectedConditions;
+
+public class RegistrationComponent extends BaseComponent {
+
+    @FindBy(xpath = "//form[contains(@action, 'register')]//input[@name='name']")
+    private WebElement nameTextField;
+
+    @FindBy(xpath = "//form[contains(@action, 'register')]//input[@name='email']")
+    private WebElement emailTextField;
+
+    @FindBy(xpath = "//form[contains(@action, 'register')]//input[@name='password']")
+    private WebElement passwordTextField;
+
+    @FindBy(xpath = "//form[contains(@action, 'register')]//input[@name='act_register_now']")
+    private WebElement registerButton;
+
+    @FindBy(xpath = "//form[contains(@action, 'register')]/p")
+    private WebElement errorMessageLabel;
+
+    public RegistrationComponent(WebDriver driver) {
+        super(driver);
+    }
+
+    public void fillNameTextField(final String name) {
+        wait.until(ExpectedConditions.visibilityOf(nameTextField)).sendKeys(name);
+    }
+
+    public void fillEmailTextField(final String email) {
+        wait.until(ExpectedConditions.visibilityOf(emailTextField)).sendKeys(email);
+    }
+
+    public void fillPasswordTextField(final String password) {
+        wait.until(ExpectedConditions.visibilityOf(passwordTextField)).sendKeys(password);
+    }
+
+    public void clickRegisterButton() {
+        wait.until(ExpectedConditions.elementToBeClickable(registerButton)).click();
+    }
+
+    public String getErrorMessageLabelText() {
+        return wait.until(ExpectedConditions.visibilityOf(errorMessageLabel)).getText();
+    }
+}
